refactor(ast): split ast.go and pool.go god-files by responsibility
Sprint E1+E2 from 2026-04-21 architect review — mechanical refactor, zero
behavior change, all tests pass.

Before PR (ast.go = 2463 lines, pool.go = 2030+ lines) → After:

  ast.go                        2463 →  178 lines
  ast_statements.go (new)               778 lines — Statement types
  ast_expressions.go (new)              773 lines — Expression types
  ast_clauses.go (new)                  791 lines — Clause / support types

  pool.go                       2030 →  869 lines — pool decls + entry points
  pool_statement_release.go (new)       766 lines — Put*Statement drivers
  pool_expression_release.go (new)      781 lines — PutExpression + helpers

All declarations preserved byte-for-byte. Cross-file references within the
same ast package resolve cleanly. Every test (go test -race ./...) still
passes. This unblocks contributor onboarding and prevents further drift:
pool.go was growing +500 lines in the last review cycle.

Also includes docs/ARCHITECT_REVIEW_2026-04-21.md — the round-2 architect
review document that drove this entire sprint series.
